Move Common/GraphicsContext.h to Common/GPU/GraphicsContext.h
