+++ Common/CMakeLists.txt
@@ -134,6 +134,7 @@ add_library(Common STATIC
 	File/DirListing.h
 	File/FileDescriptor.cpp
 	File/FileDescriptor.h
+	GPU/GraphicsContext.h
 	GPU/DataFormat.h
 	GPU/MiscTypes.h
 	GPU/GPUBackendCommon.cpp
