+++ Common/Common.vcxproj
@@ -429,6 +429,7 @@
     <ClInclude Include="GPU\D3D11\D3D11Loader.h" />
     <ClInclude Include="GPU\DataFormat.h" />
     <ClInclude Include="GPU\GPUBackendCommon.h" />
+    <ClInclude Include="GPU\GraphicsContext.h" />
     <ClInclude Include="GPU\MiscTypes.h" />
     <ClInclude Include="GPU\OpenGL\DataFormatGL.h" />
     <ClInclude Include="GPU\OpenGL\gl3stub.h" />
@@ -525,7 +526,6 @@
     <ClInclude Include="DbgNew.h" />
     <ClInclude Include="ExceptionHandlerSetup.h" />
     <ClInclude Include="GhidraClient.h" />
-    <ClInclude Include="GraphicsContext.h" />
     <ClInclude Include="Log.h" />
     <ClInclude Include="Log\LogManager.h" />
     <ClInclude Include="MachineContext.h" />
@@ -1039,6 +1039,7 @@
     <Text Include="..\ext\at3_standalone\README.txt" />
     <Text Include="..\ext\imgui\CMakeLists.txt" />
     <Text Include="..\ext\libpng17\CMakeLists.txt" />
+    <Text Include="CMakeLists.txt" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ext\cpu_features.vcxproj">
