+++ UI/UI.vcxproj.filters
@@ -150,6 +150,7 @@
     <ClCompile Include="GameBrowser.cpp">
       <Filter>Views</Filter>
     </ClCompile>
+    <ClCompile Include="LoadStateConfirmScreen.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -300,6 +301,7 @@
     <ClInclude Include="GameBrowser.h">
       <Filter>Views</Filter>
     </ClInclude>
+    <ClInclude Include="LoadStateConfirmScreen.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Screens">
@@ -311,5 +313,8 @@
     <Filter Include="ImDebugger">
       <UniqueIdentifier>{fda6bc55-1386-4650-a274-44fac9605ea3}</UniqueIdentifier>
     </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <Text Include="CMakeLists.txt" />
   </ItemGroup>
 </Project>
