+++ android/jni/AndroidJavaGLContext.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Common/GraphicsContext.h"
+#include "Common/GPU/GraphicsContext.h"
 #include "Common/GPU/OpenGL/GLRenderManager.h"
 #include "Common/GPU/thin3d_create.h"
