+++ android/jni/AndroidVulkanContext.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Common/GraphicsContext.h"
+#include "Common/GPU/GraphicsContext.h"
 #include "Common/GPU/thin3d.h"
 
 #include <android/native_window_jni.h>
